requirements.txt
openai==0.25.0
PyYAML==6.0
